Made Changes - also got rid of "Normal Button"
