+++ packages/react-app/src/components/InstructionsModal/base.tsx
@@ -1,11 +1,28 @@
 import React from "react";
 import { StatefulTabs, Tab } from "baseui/tabs";
-import tabStyle from "../../components/TabStyle";
+import * as colors from "../../constants/colors";
 import Modal from "../../components/Modal";
 import {
   OsString,
   InstructionList as IInstructionList,
 } from "../InstructionList/types";
+
+const tabStyle = ({ $active, $disabled, $theme }: any) => ({
+  outlineColor: colors.textColorWhite,
+  borderBottomColor: "#4d90e6",
+  fontWeight: 600,
+  color: $active ? colors.textColorWhite : "#d6d6d6",
+  ":hover": $disabled
+    ? {}
+    : {
+        color: colors.textColorWhite,
+      },
+  ":focus": $disabled
+    ? {}
+    : {
+        color: colors.textColorWhite,
+      },
+});
 
 interface TabsProps {
   InstructionList: IInstructionList;
