+++ packages/react-app/src/components/Options/styles.module.css
@@ -95,7 +95,7 @@
 }
 
 .marker_ubiquitous {
-  right: -0px;
+  right: 0px;
 }
 
 .caption {
